賣家結帳/退貨程式調整
1. 賣家結帳: 每筆品項的結算從加法修正為減法
2. 賣家結帳: 佣金/保險費支援小數點運算
3. 賣家結帳/退貨的dot補上拍賣公司的postfix(_G)

+++ Library/Library/InternetLibrary.cs
@@ -152,15 +152,15 @@ namespace InternetLibrary
                 CardID = IfDealedServiceFee = PostID = IfNDealedServiceFee = ContractID = "";
         }
 
-        public void parseDealedFee(out int pictureFee, out int insuranceFeeP, out int serviceFeeP)
+        public void parseDealedFee(out int pictureFee, out double insuranceFeeP, out double serviceFeeP)
         {
             pictureFee = 0;
-            insuranceFeeP = 0;
-            serviceFeeP = 0;
+            insuranceFeeP = 0.0;
+            serviceFeeP = 0.0;
 
             int.TryParse(IfDealedPictureFee, out pictureFee);
-            int.TryParse(IfDealedInsuranceFee, out insuranceFeeP);
-            int.TryParse(IfDealedServiceFee, out serviceFeeP);
+            double.TryParse(IfDealedInsuranceFee, out insuranceFeeP);
+            double.TryParse(IfDealedServiceFee, out serviceFeeP);
         }
 
         public void parseOtherFee(out int otherFee)
